Update team.md, fix icon, add team value
Fix icon as it is moved to branding folder, add team value "Raccoon is an animal with their own territory, so are we. We are proficient in our own field -- front-end developers, back-end developer, designer, team leader, tester, etc. But we are more than that. We value communication and unity. We are collaborator that scanvenge for treasures together. We value trust, courage, and creativity. Just as in our name, we are productive raccoon with productivity. The special hand of raccoon is gonna to create a better world for you."