added search function and display recipe cards on page
To display the recipe cards on the page, this file assigns the attributes to the recipe card element in the shadow DOM. Take a look at the constructor in RecipeCard.js to see how a card is created. This may need to be changed to increase the quality of the code.